Integrate CDCT into the regular ODS workflow Pacts are no longer committed to the repository. They are uploaded as an artifact instead. No separate CDCT workflow is required anymore. Reduces the number of docker-compose files by introducing additional stages in dockerfiles and overriding docker-compose files for testing.
