--- .github/workflows/cdct.yml
@@ -1,48 +1,0 @@
-name: Consumer-Driven Contract Test
-
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-    branches:
-      - main
-  workflow_dispatch:
-
-jobs:
-  cdct:
-    name: Consumer- and Provider-Side Testing
-    runs-on: ubuntu-18.04
-    
-    steps:
-      - uses: actions/checkout@v2
-
-      - name: Build Consumer-Side Test for UI
-        run: |
-          docker-compose -f ui/pact/docker-compose.consumer.yml build
-
-      - name: Build Consumer-Side Test for Notification
-        run: |
-          docker-compose -f notification/pact/docker-compose.consumer.yml build
-
-      - name: Run Consumer-Side Test for UI
-        run: |
-          docker-compose -f ui/pact/docker-compose.consumer.yml up
-
-      - name: Run Consumer-Side Test for Notification
-        run: |
-          docker-compose -f notification/pact/docker-compose.consumer.yml up
-      
-      - name: Commit latest contracts
-        uses: stefanzweifel/git-auto-commit-action@v4
-        with:
-          commit_message: Update auto-generated contracts
-          file_pattern: pacts/*.json
-      
-      - name: Build Provider-Side Test
-        run: |
-          docker-compose -f docker-compose.provider.yml build
-
-      - name: Run Provider-Side Test
-        run: |
-          docker-compose -f docker-compose.provider.yml up --abort-on-container-exit --exit-code-from pipeline-provider
