+++ .gitignore
@@ -1,4 +1,4 @@
 .idea/
 *.iml
 
-pacts/logs
+pacts/
