+++ docker-compose.consumer.yml
@@ -1,0 +1,15 @@
+version: '3.7'
+
+services:
+  ui:
+    build:
+      target: base
+    command: npm run test:consumer
+    volumes:
+      - ./pacts:/pacts
+  notification:
+    build:
+      target: base
+    command: npm run test:consumer
+    volumes:
+      - ./pacts:/pacts
