+++ docker-compose.test.yml
@@ -1,0 +1,15 @@
+version: '3.7'
+
+services:
+  ui:
+    build:
+      target: base
+    command: npm run test
+  notification:
+    build:
+      target: base
+    command: npm run test
+  pipeline:
+    build:
+      target: base
+    command: npm run test
