+++ notification/package.json
@@ -10,7 +10,7 @@
     "lint": "eslint src --fix",
     "lint-ci": "eslint src --max-warnings=0",
     "test": "jest --colors --verbose",
-    "test:pact-consumer": "jest --config ./pact/consumer.jest.config.js --colors --verbose"
+    "test:consumer": "jest --config consumer.jest.config.js --colors --verbose"
   },
   "dependencies": {
     "@jvalue/node-dry-amqp": "0.1.2",
