+++ pipeline/Dockerfile
@@ -1,7 +1,14 @@
-#-------------------------------------------------------#
-# First stage: image to build and test node application #
-#-------------------------------------------------------#
-FROM node:16-alpine as build-stage
+#----------------------------------------------------------#
+# First stage: base image for further building and testing #
+#----------------------------------------------------------#
+FROM node:16-alpine as base
+
+# install dependencies for pact (https://docs.pact.io/docker/)
+RUN apk add --no-cache --virtual build-dependencies build-base
+RUN apk --no-cache add ca-certificates wget bash \
+  && wget -q -O /etc/apk/keys/sgerrand.rsa.pub https://alpine-pkgs.sgerrand.com/sgerrand.rsa.pub \
+  && wget https://github.com/sgerrand/alpine-pkg-glibc/releases/download/2.29-r0/glibc-2.29-r0.apk \
+  && apk add glibc-2.29-r0.apk
 
 WORKDIR /build
 
@@ -14,23 +21,28 @@ RUN npm ci
 # copy rest of the files
 COPY ./src ./src
 COPY ./tsconfig.json ./
-COPY ./.eslintrc.js ./
-COPY ./jest.config.js ./
+COPY ./*.js ./
 
-# build and test
+# lint project
 RUN npm run lint-ci
-RUN npm run transpile
-RUN npm run test
 
-#---------------------------------------------#
-# Second stage: image to run node application #
-#---------------------------------------------#
+#-----------------------------------------------#
+# Second stage: image to build node application #
+#-----------------------------------------------#
+FROM base as build
+
+# build
+RUN npm run transpile
+
+#--------------------------------------------#
+# Third stage: image to run node application #
+#--------------------------------------------#
 FROM node:16-alpine
 
 WORKDIR /app
 
-COPY --from=build-stage /build/dist/ ./dist/
-COPY --from=build-stage /build/package*.json ./
+COPY --from=build /build/dist/ ./dist/
+COPY --from=build /build/package*.json ./
 
 RUN npm ci --only=production
 EXPOSE 8080
