+++ pipeline/package.json
@@ -10,7 +10,7 @@
     "lint": "eslint src --fix",
     "lint-ci": "eslint src --max-warnings=0",
     "test": "jest --colors --verbose",
-    "test:pact-provider": "jest --config ./pact/provider.jest.config.js --colors --verbose"
+    "test:provider": "jest --config provider.jest.config.js --colors --verbose"
   },
   "dependencies": {
     "@jvalue/node-dry-amqp": "0.1.2",
