--- pipeline/pact/Dockerfile.provider
@@ -1,25 +1,0 @@
-FROM node:15-alpine
-
-# install dependencies for pact (https://docs.pact.io/docker/)
-RUN apk add --no-cache --virtual build-dependencies build-base
-RUN apk --no-cache add ca-certificates wget bash \
-  && wget -q -O /etc/apk/keys/sgerrand.rsa.pub https://alpine-pkgs.sgerrand.com/sgerrand.rsa.pub \
-  && wget https://github.com/sgerrand/alpine-pkg-glibc/releases/download/2.29-r0/glibc-2.29-r0.apk \
-  && apk add glibc-2.29-r0.apk
-
-WORKDIR /app/pipeline
-
-# Copy package*.json files first in order to make best use of docker layer caching
-COPY ./pipeline/package*.json ./
-
-# npm clean slate install to get reproducible builds and quicker installs
-RUN npm ci
-
-# copy rest of the files
-COPY ./pipeline/src ./src
-COPY ./pipeline/*.js ./
-COPY ./pipeline/*.json ./
-COPY ./pipeline/pact ./pact
-
-# run provider tests
-CMD ["npm", "run", "test:pact-provider"]
