+++ ui/package.json
@@ -7,7 +7,7 @@
     "lint": "eslint src --fix",
     "lint-ci": "eslint src --max-warnings=0",
     "test": "jest --colors --verbose --passWithNoTests",
-    "test:pact-consumer": "jest --config ./pact/consumer.jest.config.js --colors --verbose"
+    "test:consumer": "jest --config consumer.jest.config.js --colors --verbose"
   },
   "dependencies": {
     "axios": "^0.21.1",
